requirements.txt
openai>=1.0.0
pytesseract
Pillow
python-dotenv
fastapi
uvicorn
pydantic
tqdm
aiofiles
httpx
loguru
opencv-python
pdf2image
python-multipart
pytesseract
easyocr
paddleocr
paddlepaddle==2.5.0  # oppure versione CPU compatibile
Pillow
